Survey info

Get information about all DEV Qualtrics surveys, including number of observations in raw data and date of most recent response

get_survey_info(append_raw_data = FALSE)

Arguments

append_raw_data
A logical value indicating whether raw survey data should be included as a column in the output

Value

A tibble with the following columns:

id (unique survey identifier)

name (survey name)

lastModified (date of most recent response)

n_obs (number of observations in raw survey data)

If append_raw_data is set to TRUE, output includes an additional list column, survey_data.